Update README.md
This is my first full power bi project that I did. It was tremendously challenging, but fulfilling since I learned so much during this phase. And it was this particular project that instilled the confidence in my ability to use the application, and encouraged me to learn more. Its interesting that the thing that i most struggle with, was to do most with the visuals. The cleaning of data, creating relationships as well as using DAX, not as challenging. Its the presentation that I will be working to harness. 

also, in my next ones, i will be able to include insight report.